moved neuron data away from a seperate file, fixed some cartesian conversions
